v.7

+++ iBot/files_activities.py
@@ -3,6 +3,8 @@ import shutil
 import time
 import requests
 import PyPDF2
+import PIL
+from shutil import copyfile
 
 
 def convert_bytes(num):
@@ -15,12 +17,16 @@ def convert_bytes(num):
 class File:
 
     def __init__(self, path):
-
-        self.exists = os.path.isfile(path)
         self.path = path
-        self.byteSize = convert_bytes(os.stat(path).st_size)
-        self.creationTime = time.ctime(os.path.getctime(path))
-        self.modificationTime = time.ctime(os.path.getmtime(path))
+        self.exists = os.path.isfile(self.path)
+        if self.exists:
+            self.byteSize = convert_bytes(os.stat(path).st_size)
+            self.creationTime = time.ctime(os.path.getctime(path))
+            self.modificationTime = time.ctime(os.path.getmtime(path))
+        else:
+            self.byteSize = None
+            self.creationTime = None
+            self.modificationTime = None
 
     def rename(self, new_file_name):
         old_path = self.path
@@ -54,35 +60,40 @@ class File:
             elif "/" in self.path:
                 new_path = new_location + "/" + "(" + str(uuid.uuid4())[:8] + ") " + name
             os.rename(self.path, new_path)
-            self.path = new_path
+
+        self.path = new_path
 
     def remove(self):
         if os.path.isfile(self.path):
             os.remove(self.path)
 
-    def copy(self, new_path):
-        from shutil import copyfile
-        copyfile(self.path, new_path)
+    def copy(self, new_location):
 
-    def WaitFor(self):
+        import uuid
+        if "\\" in self.path:
+            name = self.path.split("\\")[-1]
+            new_path = new_location + "\\" + name
+        elif "/" in self.path:
+            name = self.path.split("/")[-1]
+            new_path = new_location + "/" + name
+        if os.path.exists(self.path):
+            if not os.path.exists(new_path):
+                copyfile(self.path, new_path)
+        elif os.path.exists(new_path):
+            if "\\" in self.path:
+                new_path = new_location + "\\" + "(" + str(uuid.uuid4())[:8] + ") " + name
+            elif "/" in self.path:
+                new_path = new_location + "/" + "(" + str(uuid.uuid4())[:8] + ") " + name
+            copyfile(self.path, new_path)
+
+    def waitFor(self):
         from time import sleep
-        while not os.path.exists(self.path):
+        while not os.path.isfile(self.path):
             sleep(1)
-
-    @staticmethod
-    def download(url, path, name=None):
-        if not name:
-            filename = path + "/" + name
-        else:
-            filename = path + "/" + str(url).split('/')[-1]
-        r = requests.get(url, allow_redirects=True)
-        open(filename, 'wb').write(r.content)
-
-        return filename
+        self.__init__(self.path)
 
 
 class PDF(File):
-
     def __init__(self, path):
         File.__init__(self, path)
 
@@ -114,37 +125,35 @@ class PDF(File):
 class Image(File):
     def __init__(self, path):
         File.__init__(self, path)
-
-        self.size = str(Image.open(path).size)
-        self.format = Image.open(path).format
+        self.size = self.open().size
+        self.format = self.open().format
 
     def open(self):
-        im = Image.open(self.path)
-        return im.show()
+        im = PIL.Image.open(self.path)
+        return im
 
     def rotate(self, angle):
-        im = Image.open(self.path)
+        im = self.open()
         return im.rotate(angle, expand=True).save(self.path)
 
     def resize(self, size):
-        im = Image.open(self.path)
+        im = self.open()
         return im.resize(size).save(self.path)
 
-    def crop(self, path, box=None):
-        im = Image.open(path)
-        return im.crop(box).save(path)
+    def crop(self, box=None):
+        im = self.open()
+        return im.crop(box).save(self.path)
 
     # Mirrors an image with a given path from left to right.
 
     def mirrorH(self):
-        im = Image.open(self.path)
-        return im.transpose(Image.FLIP_LEFT_RIGHT).save(self.path)
+        im = self.open()
+        return im.transpose(PIL.Image.FLIP_LEFT_RIGHT).save(self.path)
 
     #    Mirrors an image with a given path from top to bottom.
-
     def mirrorV(self):
-        im = Image.open(self.path)
-        return im.transpose(Image.FLIP_TOP_BOTTOM).save(self.path)
+        im = self.open()
+        return im.transpose(PIL.Image.FLIP_TOP_BOTTOM).save(self.path)
 
 
 class Folder:
@@ -235,3 +244,12 @@ class Folder:
         if not os.path.exists(self.path):
             os.makedirs(self.path)
         return self.path
+
+    def downloadFile(self, url, name=None):
+        if not name:
+            filename = self.path + "/" + name
+        else:
+            filename = self.path + "/" + str(url).split('/')[-1]
+        r = requests.get(url, allow_redirects=True)
+        open(filename, 'wb').write(r.content)
+        return filename
